tests/requirements.txt
pytest>=7.0.0
requests>=2.28.0
pyyaml>=6.0
kubernetes>=24.0.0
pytest-html>=3.1.0
pytest-cov>=4.0.0
locust>=2.14.0